Validar funcionalidades de la pagina dependiendo de los permisos

+++ app/Http/Controllers/PermissionController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
 
 class PermissionController extends Controller
@@ -14,6 +15,7 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('permission_index'), 403);
         //
         $permissions = Permission::all();
 
@@ -28,6 +30,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('permission_create'), 403);
         //
         return view('pages.permissions.create');
     }
@@ -67,6 +70,7 @@ class PermissionController extends Controller
     public function edit(Permission $permission)
     {
         //
+        abort_if(Gate::denies('permission_edit'), 403);
         return view('pages.permissions.edit', compact('permission'));
     }
 
@@ -94,6 +98,7 @@ class PermissionController extends Controller
     public function destroy(Permission $permission)
     {
         //
+        abort_if(Gate::denies('permission_destroy'), 403);
         $permission->delete();
 
         return redirect()->route('permission.index');
